Edit med UI, link to backend, other fixes
