+++ src/App.tsx
@@ -2,7 +2,7 @@ import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 import WelcomePage from './components/WelcomePage'
 import CreateAccount from './components/CreateAcct'
 import HomePage from './components/HomePage.tsx'
-import EditMedicationsPage from './components/EditMed.tsx'
+import EditMed from './components/EditMed.tsx'
 import CreateMed from './components/CreateMed.tsx'
 import './App.css';
 
@@ -14,7 +14,7 @@ function App() {
               <Route path="/create-account" element={<CreateAccount/>}/>
               <Route path="/home" element={<HomePage/>} />
               <Route path="/create-med" element={<CreateMed/>}/>
-              <Route path="/edit-medications" element={<EditMedicationsPage />} />
+              <Route path="/edit-med" element={<EditMed/>}/>
       </Routes>
     </Router>
