+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/createacct.tsx","./src/components/createmed.tsx","./src/components/editmed.tsx","./src/components/homepage.tsx","./src/components/welcomepage.tsx","./src/scripts/fcm.tsx","./src/scripts/firebase-init.tsx","./src/scripts/medcalls.tsx","./src/scripts/userauth.tsx","./src/types/types.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/createacct.tsx","./src/components/createmed.tsx","./src/components/editmed.tsx","./src/components/homepage.tsx","./src/components/medmenu.tsx","./src/components/welcomepage.tsx","./src/scripts/fcm.tsx","./src/scripts/firebase-init.tsx","./src/scripts/medcalls.tsx","./src/scripts/userauth.tsx","./src/types/types.ts"],"version":"5.6.3"}
